test(release-3.0): isolate daemon recovery regressions
