+++ tests/test_guard_daemon_manager.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import hashlib
 import json
 import os
-import signal
 import stat
 import subprocess
 import sys
@@ -37,15 +36,6 @@ class _WindowsOSProxy:
     """Expose Windows branching without mutating process-wide ``os.name``."""
 
     name = "nt"
-
-    def __getattr__(self, name: str):
-        return getattr(os, name)
-
-
-class _PosixOSProxy:
-    """Expose POSIX branching without mutating process-wide ``os.name``."""
-
-    name = "posix"
 
     def __getattr__(self, name: str):
         return getattr(os, name)
@@ -2740,47 +2730,6 @@ def test_authenticated_state_with_proven_foreign_recycled_pid_is_tombstoned(tmp_
 
     assert daemon_manager_module.retire_all_guard_daemons_for_home(guard_home) == []
     assert state_clears == [62_222]
-
-
-def test_posix_daemon_retirement_waits_for_sigkill_to_finish(monkeypatch) -> None:
-    pid = 62_223
-    signals: list[int] = []
-    waits = iter((False, True))
-    sigkill = getattr(signal, "SIGKILL", 9)
-
-    monkeypatch.setattr(daemon_manager_module, "os", _PosixOSProxy())
-    monkeypatch.setattr(daemon_manager_module.signal, "SIGKILL", sigkill, raising=False)
-    monkeypatch.setattr(daemon_manager_module, "_guard_daemon_pid_is_proven_dead", lambda _pid: False)
-    monkeypatch.setattr(daemon_manager_module, "_guard_daemon_pid_matches_command", lambda *_args: True)
-    monkeypatch.setattr(
-        daemon_manager_module,
-        "_wait_for_guard_daemon_pid_death",
-        lambda _pid: next(waits),
-    )
-    monkeypatch.setattr(daemon_manager_module.os, "kill", lambda _pid, sig: signals.append(sig))
-
-    assert daemon_manager_module._retire_guard_daemon_pid(pid) is True
-    assert signals == [signal.SIGTERM, sigkill]
-
-
-@pytest.mark.parametrize("failing_signal", (signal.SIGTERM, getattr(signal, "SIGKILL", 9)))
-def test_posix_daemon_retirement_does_not_accept_signal_permission_error(monkeypatch, failing_signal) -> None:
-    pid = 62_224
-    sigkill = getattr(signal, "SIGKILL", 9)
-
-    monkeypatch.setattr(daemon_manager_module, "os", _PosixOSProxy())
-    monkeypatch.setattr(daemon_manager_module.signal, "SIGKILL", sigkill, raising=False)
-    monkeypatch.setattr(daemon_manager_module, "_guard_daemon_pid_is_proven_dead", lambda _pid: False)
-    monkeypatch.setattr(daemon_manager_module, "_guard_daemon_pid_matches_command", lambda *_args: True)
-    monkeypatch.setattr(daemon_manager_module, "_wait_for_guard_daemon_pid_death", lambda _pid: False)
-
-    def deny_signal(_pid: int, sent_signal: int) -> None:
-        if sent_signal == failing_signal:
-            raise PermissionError("signal denied")
-
-    monkeypatch.setattr(daemon_manager_module.os, "kill", deny_signal)
-
-    assert daemon_manager_module._retire_guard_daemon_pid(pid) is False
 
 
 def test_malformed_windows_lifecycle_records_are_quarantined_after_two_empty_inventories(
